Corrected uses of Whitelist to Safelist
And introduced a (deprecated) compatibility shim.

Tested with japicmp binary and source compatibility report, and checked binary compat with some test utilities (local testing).

Many thanks to @drei01 (#1408), @nkiesel (#1423) and others for highlighting the issue and for PRs which this commit refines.
