+++ CHANGES
@@ -5,6 +5,10 @@ jsoup changelog
 
   * Change: updated the minimum Android API level from 8 to 10.
 
+  * Improvement: renamed the Whitelist class to Safelist, with the goal of more inclusive language. A shim is provided
+    for backwards compatibility (source and binary). This shim is marked as deprecated and will be removed in the
+    jsoup 1.15.1 release.
+
   * Improvement: added support for loading and parsing gzipped HTML files in Jsoup.parse(File in, charset, baseUri).
 
   * Improvement: reduced thread content in HttpConnection and Document.
@@ -458,12 +462,12 @@ jsoup changelog
  * Added the new selector :containsData(), to find elements that hold data, like script and style tags.
 
  * Changed Jsoup.isValid(bodyHtml) to validate that the input contains only body HTML that is safe according to the
-   whitelist, and does not include HTML errors. And in the Jsoup.Cleaner.isValid(Document) method, make sure the doc
+   safelist, and does not include HTML errors. And in the Jsoup.Cleaner.isValid(Document) method, make sure the doc
    only includes body HTML.
    <https://github.com/jhy/jsoup/issues/245>
    <https://github.com/jhy/jsoup/issues/632>
 
- * In Whitelists, validate that a removed protocol exists before removing said protocol.
+ * In Safelists, validate that a removed protocol exists before removing said protocol.
 
  * Allow the Jsoup.Connect thread to be interrupted when reading the input stream; helps when reading from a long stream
    of data that doesn't read timeout.
@@ -658,10 +662,10 @@ jsoup changelog
    or your JDK doesn't support SNI.
    <https://github.com/jhy/jsoup/pull/343>
 
- * Added ability to further tweak the canned Cleaner Whitelists by removing existing settings.
+ * Added ability to further tweak the canned Cleaner Safelists by removing existing settings.
    <https://github.com/jhy/jsoup/pull/449>
 
- * Added option in Cleaner Whitelist to allow linking to in-page anchors (#)
+ * Added option in Cleaner Safelist to allow linking to in-page anchors (#)
    <https://github.com/jhy/jsoup/pull/441>
 
  * Use a lowercase doctype tag for HTML5 documents.
@@ -734,7 +738,7 @@ jsoup changelog
  * If pretty-print is disabled, don't trim outer whitespace in Element.html()
    <https://github.com/jhy/jsoup/issues/368>
 
- * In the HTML Cleaner, allow span tags in the basic whitelist, and span and div tags in the relaxed whitelist.
+ * In the HTML Cleaner, allow span tags in the basic safelist, and span and div tags in the relaxed safelist.
 
   * Added Element.cssSelector(), which returns a unique CSS selector/path for an element.
     <https://github.com/jhy/jsoup/pull/459>
@@ -769,7 +773,7 @@ jsoup changelog
  * Added support for 'application/*+xml' mimetypes.
    <https://github.com/jhy/jsoup/pull/444>
 
- * Fixed support for allowing script tags in cleaner whitelists.
+ * Fixed support for allowing script tags in cleaner Safelists.
    <https://github.com/jhy/jsoup/issues/299>
    <https://github.com/jhy/jsoup/issues/388>
 
@@ -833,7 +837,7 @@ jsoup changelog
  * Introduced Parser.parseXmlFragment(), to allow easy parsing of XML fragments.
    <https://github.com/jhy/jsoup/issues/279>
 
- * Allow Whitelist test methods to be extended
+ * Allow Safelist test methods to be extended
    <https://github.com/jhy/jsoup/issues/85>
 
  * Added Document.OutputSettings.outline mode, to aid HTML debugging by printing out in outline mode, similar to
@@ -925,13 +929,13 @@ jsoup changelog
  * Fixed NPE when HTML fragment parsing a <style> tag
    <https://github.com/jhy/jsoup/issues/189>
 
- * Fixed issue with :all pseudo-tag in HTML sanitizer when cleaning tags previously defined in whitelist
+ * Fixed issue with :all pseudo-tag in HTML sanitizer when cleaning tags previously defined in safelist
    <https://github.com/jhy/jsoup/issues/156>
 
  * Fixed NPE in Parser.parseFragment() when context parameter is null.
    <https://github.com/jhy/jsoup/issues/195>
 
- * In HTML whitelists, when defining allowed attributes for a tag, automatically add the tag to the allowed list.
+ * In HTML Safelists, when defining allowed attributes for a tag, automatically add the tag to the allowed list.
 
 *** Release 1.6.2 [2012-Mar-27]
  * Added a simplified XML parsing mode, which can usefully parse valid and invalid XML, but does not enforce any HTML
@@ -950,7 +954,7 @@ jsoup changelog
  * Updated jsoup.connect so that when requests made as POSTs are redirected, the redirect is followed as a GET.
    <https://github.com/jhy/jsoup/issues/120>
 
- * Updated the Cleaner and whitelists to optionally preserve related links in elements, instead of converting them
+ * Updated the Cleaner and Safelists to optionally preserve related links in elements, instead of converting them
    to absolute links.
 
  * Updated the Cleaner to support custom allowed protocols such as "cid:" and "data:".
@@ -1255,7 +1259,7 @@ jsoup changelog
     prepend(String), append(String); bulk methods for corresponding
     methods in Element.
 
- * New feature: Jsoup.isValid(html, whitelist) method for user input
+ * New feature: Jsoup.isValid(html, safelist) method for user input
     form validation.
  
  * Improved Elements.attr(String) to find first matching element
