+++ src/test/java/org/jsoup/safety/CompatibilityTests.java
@@ -1,0 +1,99 @@
+package org.jsoup.safety;
+
+import org.jsoup.Jsoup;
+import org.jsoup.MultiLocaleExtension;
+import org.jsoup.TextUtil;
+import org.jsoup.nodes.Document;
+import org.jsoup.nodes.Entities;
+import org.junit.jupiter.api.Test;
+
+import java.util.Locale;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+/**
+ Tests for the deprecated {@link org.jsoup.safety.Whitelist} class source compatibility. Will be removed in
+ <code>v.1.15.1</code>. No net new tests here so safe to blow up.
+ */
+public class CompatibilityTests {
+    @Test
+    public void resolvesRelativeLinks() {
+        String html = "<a href='/foo'>Link</a><img src='/bar'>";
+        String clean = Jsoup.clean(html, "http://example.com/", Whitelist.basicWithImages());
+        assertEquals("<a href=\"http://example.com/foo\" rel=\"nofollow\">Link</a>\n<img src=\"http://example.com/bar\">", clean);
+    }
+
+    @Test
+    public void testDropsUnknownTags() {
+        String h = "<p><custom foo=true>Test</custom></p>";
+        String cleanHtml = Jsoup.clean(h, Whitelist.relaxed());
+        assertEquals("<p>Test</p>", cleanHtml);
+    }
+
+    @Test
+    public void preservesRelativeLinksIfConfigured() {
+        String html = "<a href='/foo'>Link</a><img src='/bar'> <img src='javascript:alert()'>";
+        String clean = Jsoup.clean(html, "http://example.com/", Whitelist.basicWithImages().preserveRelativeLinks(true));
+        assertEquals("<a href=\"/foo\" rel=\"nofollow\">Link</a>\n<img src=\"/bar\"> \n<img>", clean);
+    }
+
+    @Test
+    public void handlesCustomProtocols() {
+        String html = "<img src='cid:12345' /> <img src='data:gzzt' />";
+        String dropped = Jsoup.clean(html, Whitelist.basicWithImages());
+        assertEquals("<img> \n<img>", dropped);
+
+        String preserved = Jsoup.clean(html, Whitelist.basicWithImages().addProtocols("img", "src", "cid", "data"));
+        assertEquals("<img src=\"cid:12345\"> \n<img src=\"data:gzzt\">", preserved);
+    }
+
+    @Test
+    public void handlesFramesets() {
+        String dirty = "<html><head><script></script><noscript></noscript></head><frameset><frame src=\"foo\" /><frame src=\"foo\" /></frameset></html>";
+        String clean = Jsoup.clean(dirty, Whitelist.basic());
+        assertEquals("", clean); // nothing good can come out of that
+
+        Document dirtyDoc = Jsoup.parse(dirty);
+        Document cleanDoc = new Cleaner(Whitelist.basic()).clean(dirtyDoc);
+        assertNotNull(cleanDoc);
+        assertEquals(0, cleanDoc.body().childNodeSize());
+    }
+
+    @Test
+    public void supplyOutputSettings() {
+        // test that one can override the default document output settings
+        Document.OutputSettings os = new Document.OutputSettings();
+        os.prettyPrint(false);
+        os.escapeMode(Entities.EscapeMode.extended);
+        os.charset("ascii");
+
+        String html = "<div><p>&bernou;</p></div>";
+        String customOut = Jsoup.clean(html, "http://foo.com/", Whitelist.relaxed(), os);
+        String defaultOut = Jsoup.clean(html, "http://foo.com/", Whitelist.relaxed());
+        assertNotSame(defaultOut, customOut);
+
+        assertEquals("<div><p>&Bscr;</p></div>", customOut); // entities now prefers shorted names if aliased
+        assertEquals("<div>\n" +
+            " <p>ℬ</p>\n" +
+            "</div>", defaultOut);
+
+        os.charset("ASCII");
+        os.escapeMode(Entities.EscapeMode.base);
+        String customOut2 = Jsoup.clean(html, "http://foo.com/", Whitelist.relaxed(), os);
+        assertEquals("<div><p>&#x212c;</p></div>", customOut2);
+    }
+
+    @MultiLocaleExtension.MultiLocaleTest
+    public void safeListedProtocolShouldBeRetained(Locale locale) {
+        Locale.setDefault(locale);
+
+        Whitelist safelist = Whitelist.none()
+            .addTags("a")
+            .addAttributes("a", "href")
+            .addProtocols("a", "href", "something");
+
+        String cleanHtml = Jsoup.clean("<a href=\"SOMETHING://x\"></a>", safelist);
+
+        assertEquals("<a href=\"SOMETHING://x\"></a>", TextUtil.stripNewlines(cleanHtml));
+    }
+}
